skills/pdfkit-py/scripts/requirements.txt
# pdfkit lite - core Python dependencies
# Optional dependencies are installed on demand by commands

# Core dependencies
PyMuPDF>=1.23.0
pypdf>=3.0.0
Pillow>=10.0.0
reportlab>=4.0.0
numpy>=1.24.0

# PDF parsing and structure detection
pdfplumber>=0.9.0

# OCR wrapper
pytesseract>=0.3.10

# Format conversion basics
markdown>=3.4.0
python-docx>=0.8.11

# Font inspection
fonttools>=4.0.0
